refactor: StaffController 통합
