+++ backend/src/test/java/com/festago/presentation/StaffControllerTest.java
@@ -11,7 +11,11 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.festago.auth.application.StaffAuthService;
 import com.festago.auth.dto.StaffLoginRequest;
 import com.festago.auth.dto.StaffLoginResponse;
+import com.festago.entry.application.EntryService;
+import com.festago.entry.dto.TicketValidationRequest;
+import com.festago.entry.dto.TicketValidationResponse;
 import com.festago.support.CustomWebMvcTest;
+import com.festago.ticketing.domain.EntryState;
 import java.nio.charset.StandardCharsets;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
@@ -35,6 +39,9 @@ class StaffControllerTest {
     @MockBean
     StaffAuthService staffAuthService;
 
+    @MockBean
+    EntryService entryService;
+
     @Test
     void 스태프_로그인() throws Exception {
         // given
@@ -54,6 +61,27 @@ class StaffControllerTest {
             .getContentAsString(StandardCharsets.UTF_8);
         StaffLoginResponse actual = objectMapper.readValue(content, StaffLoginResponse.class);
         assertThat(actual).isEqualTo(expected);
+    }
 
+    @Test
+    void QR을_검사한다() throws Exception {
+        // given
+        TicketValidationRequest request = new TicketValidationRequest("anyCode");
+        TicketValidationResponse expected = new TicketValidationResponse(EntryState.AFTER_ENTRY);
+        given(entryService.validate(request))
+            .willReturn(expected);
+
+        // when & then
+        String content = mockMvc.perform(post("/staff/member-tickets/validation")
+                .content(objectMapper.writeValueAsString(request))
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isOk())
+            .andDo(print())
+            .andReturn()
+            .getResponse()
+            .getContentAsString(StandardCharsets.UTF_8);
+        TicketValidationResponse actual = objectMapper.readValue(content,
+            TicketValidationResponse.class);
+        assertThat(actual).isEqualTo(expected);
     }
 }
